fix loo crossval

+++ linear_svm_crossval.py
@@ -84,7 +84,7 @@ def run_svm(X_train, y_train, X_val, y_val):
 limited, full = resample_balanced('volumes.csv')
 
 loocv = ParallelKFold()
-acc = loocv.k_fold(limited.shape[0]/2, run_svm, limited.to_numpy(), full['Target_cat'].to_numpy())
+acc = loocv.k_fold(int(limited.shape[0]/2), run_svm, limited.to_numpy(), full['Target_cat'].to_numpy())
 print("linear svm leave one out cross-validation accuracy (balanced unscaled dataset):", acc)
 
 
@@ -92,5 +92,5 @@ print("linear svm leave one out cross-validation accuracy (balanced unscaled dat
 limited, full = resample_balanced('scaled_volumes.csv')
 
 loocv = ParallelKFold()
-acc = loocv.k_fold(limited.shape[0]/2, run_svm, limited.to_numpy(), full['Target_cat'].to_numpy())
+acc = loocv.k_fold(int(limited.shape[0]/2), run_svm, limited.to_numpy(), full['Target_cat'].to_numpy())
 print("linear svm leave one out cross-validation accuracy (balanced unscaled dataset):", acc)
